chore: Move fixtures to conftest.py
